Removed bug where warningshot would crash if a resolvers order was not set
WarningShot will now chdir to the application before running (default location is '.')
This allows for warningshot to be run from locations outside the app root

Additional resolver paths can be specified, by default wanringshot looks for
additional resolvers to load in ~/.warningshot/*.rb

All base resolvers are loaded as well.  This allows for custom resolvers
to be loaded at runtime that do not come with WarningShot.